Install AWS CLI via pip
At least one critical Nextstrain build (ncov) requires AWS CLI to run, but this
dependency is only available through the Docker image [1] or when users manually
install these tools. This commit should make the AWS tools available to all
users working with Nextstrain with no effort on their part.

Note, this commit installs awscli from PyPI instead of a conda channel. Support
for conda packages of awscli seems to be inconsistent, so this approach is more
likely to work in the future.

[1] https://github.com/nextstrain/docker-base/blob/master/Dockerfile#L109