Pin Snakemake version
Snakemake versions 5.27.* recently changed the minimum Python version
supported to 3.7, changed it back to 3.5, and ended up breaking DAG
construction for Python <3.7. Although there is an open PR to resolve
this issue [1], our Nextstrain pipelines are currently broken and need
to be fixed. This commit pins the latest working version of Snakemake
until the issue described above gets resolved.

[1] https://github.com/snakemake/snakemake/pull/726